# modules/services/src/main/resources/openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "LEOS SERVICE API",
    "description" : "API documentation for LEOS core services",
    "version" : "5.5.0-SNAPSHOT"
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/{proposalRef}/getExports" : {
      "get" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Get exports",
        "description" : "Retrieves all export packages for the specified proposal",
        "operationId" : "getExports",
        "parameters" : [ {
          "name" : "proposalRef",
          "in" : "path",
          "description" : "Proposal reference",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Exports retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/searchUser" : {
      "get" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Search user",
        "description" : "Searches for users in the repository by search key",
        "operationId" : "searchUser",
        "parameters" : [ {
          "name" : "searchKey",
          "in" : "query",
          "description" : "Search key",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Users retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/copyAct" : {
      "post" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Copy act",
        "description" : "Creates a copy of an existing act with all its associated documents",
        "operationId" : "copyAct",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProposalCopyRequest",
                "exampleSetFlag" : false
              },
              "exampleSetFlag" : false
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Act copied successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/{proposalRef}" : {
      "put" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Update proposal metadata",
        "description" : "Updates metadata information for the specified proposal",
        "operationId" : "updateProposalMetadata",
        "parameters" : [ {
          "name" : "proposalRef",
          "in" : "path",
          "description" : "Proposal reference",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProposalRequest",
                "exampleSetFlag" : false
              },
              "exampleSetFlag" : false
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Metadata updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Delete proposal",
        "description" : "Permanently deletes the specified proposal and all its associated documents",
        "operationId" : "deleteProposal",
        "parameters" : [ {
          "name" : "proposalRef",
          "in" : "path",
          "description" : "Proposal reference",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Proposal deleted successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/searchUsersByJobTitle" : {
      "get" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Search users by job title",
        "description" : "Retrieves users from the repository filtered by job title",
        "operationId" : "searchUsersByJobTitle",
        "parameters" : [ {
          "name" : "jobTitle",
          "in" : "query",
          "description" : "Job title",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Users retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/{legFileId}/linguistic-versions" : {
      "post" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Create linguistic versions",
        "description" : "Creates linguistic versions from a milestone for the specified leg file",
        "operationId" : "createLinguisticVersions",
        "parameters" : [ {
          "name" : "legFileId",
          "in" : "path",
          "description" : "Leg file ID",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/List",
                "exampleSetFlag" : false
              },
              "exampleSetFlag" : false
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Linguistic versions created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/createExplanatoryDocument" : {
      "post" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Create explanatory document",
        "description" : "Creates a new explanatory document with specified template, purpose and EEA relevance",
        "operationId" : "createExplanatoryDocument",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateExplanatoryDocumentRequest",
                "exampleSetFlag" : false
              },
              "exampleSetFlag" : false
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Explanatory document created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "400" : {
            "description" : "Method not supported"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/updateDocPurpose/{proposalRef}" : {
      "put" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Update proposal document purpose",
        "description" : "Updates the document purpose and EEA relevance for the specified proposal",
        "operationId" : "updateProposalDocPurpose",
        "parameters" : [ {
          "name" : "proposalRef",
          "in" : "path",
          "description" : "Proposal reference",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProposalRequest",
                "exampleSetFlag" : false
              },
              "exampleSetFlag" : false
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Document purpose updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "{proposalRef}/create-financial-statement" : {
      "post" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Create financial statement",
        "description" : "Creates a financial statement document for the specified proposal",
        "operationId" : "createFinancialStatement",
        "parameters" : [ {
          "name" : "proposalRef",
          "in" : "path",
          "description" : "Proposal reference",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Financial statement created successfully"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "{proposalRef}/delete-financial-statement/{financialStatementRef}" : {
      "delete" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Delete financial statement",
        "description" : "Deletes the specified financial statement from the proposal",
        "operationId" : "deleteFinancialStatement",
        "parameters" : [ {
          "name" : "proposalRef",
          "in" : "path",
          "description" : "Proposal reference",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        }, {
          "name" : "financialStatementRef",
          "in" : "path",
          "description" : "Financial statement reference",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Financial statement deleted successfully"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/{proposalRef}/export/download" : {
      "get" : {
        "tags" : [ "Proposal Management" ],
        "summary" : "Download export proposal",
        "description" : "Downloads the exported proposal in the specified output format",
        "operationId" : "exportProposalDownload",
        "parameters" : [ {
          "name" : "proposalRef",
          "in" : "path",
          "description" : "Proposal reference",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        }, {
          "name" : "exportOutput",
          "in" : "query",
          "description" : "Export output",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Export downloaded successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "HTTP",
        "description" : "JWT Bearer token authentication",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "basicAuth" : {
        "type" : "HTTP",
        "description" : "Basic HTTP authentication",
        "scheme" : "basic"
      }
    }
  }
}